services/ml-worker/requirements.txt
# ML/AI (torch pre-installed in NGC image)
diffusers>=0.28.0
transformers>=4.40.0
accelerate>=0.25.0
safetensors>=0.4.0
# xformers removed - not compatible with ARM64

# Celery
celery[redis]>=5.3.0

# Storage
minio>=7.2.0

# Image processing
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime-gpu>=1.16.0

# Segment Anything Model
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Monitoring
prometheus-client>=0.19.0

# Logging
structlog>=24.1.0

# HuggingFace Hub
huggingface-hub==0.21.4

# Translation
sentencepiece>=0.1.99
